Split up hooks & better organize for dev tool inspection
