+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-fetch-hooks",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "A set of react hooks to work with the fetch API and gracefully parse & deal with HTTP errors",
   "keywords": [
     "react",
