+++ src/lazy-fetch.js
@@ -1,15 +1,23 @@
-import { useState, useEffect, useCallback } from "react";
 import { isString } from "lodash";
 
-import checkStatus from "./check-status";
+import useRequestInitialState from "./request-initial-state";
+import useFetchFn from "./fetch-fn";
+import useRefreshInterval from "./refresh-interval";
+import useResetDelay from "./reset-delay";
 
 const useLazyFetch = itemToFetch => {
-	const [isFetching, setIsFetching] = useState(false);
-	const [isFetched, setIsFetched] = useState(false);
-	const [data, setData] = useState(null);
-	const [error, setError] = useState(null);
-
-	const [timerSignal, resetTimer] = useState(0);
+	const {
+		isFetching,
+		setIsFetching,
+		isFetched,
+		setIsFetched,
+		data,
+		setData,
+		error,
+		setError,
+		timerSignal,
+		resetTimer
+	} = useRequestInitialState();
 
 	let { url, resetDelay, refreshInterval, ...opts } = parseItemToFetch(
 		itemToFetch
@@ -21,72 +29,28 @@ const useLazyFetch = itemToFetch => {
 		);
 	}
 
-	const fetchFn = useCallback(() => {
-		resetTimer(0);
+	const fetchFn = useFetchFn({
+		refreshInterval,
+		resetDelay,
+		url,
+		opts,
+		resetTimer,
+		setIsFetching,
+		setIsFetched,
+		setError,
+		setData
+	});
 
-		if (url) {
-			setIsFetching(true);
-			setError(null);
-			doFetch();
-		}
+	useResetDelay({
+		resetDelay,
+		setData,
+		setError,
+		setIsFetched,
+		setIsFetching,
+		timerSignal
+	});
 
-		async function doFetch() {
-			try {
-				let response = await fetch(url, opts);
-
-				response = await checkStatus(response);
-
-				if (response.status != 204) {
-					response = await response.json();
-				} else {
-					// for 204 No Content, just return null data
-					response = null;
-				}
-
-				setData(response);
-				setIsFetching(false);
-				setIsFetched(true);
-				setError(null);
-
-				if (resetDelay) {
-					resetTimer(s => s + 1);
-				}
-
-				if (refreshInterval) {
-					resetTimer(s => s + 1);
-				}
-			} catch (ex) {
-				setIsFetching(false);
-				setError(ex);
-			}
-		}
-		// JSON.stringify: just relax, it's fine
-		// https://github.com/facebook/react/issues/14476#issuecomment-471199055
-	}, [refreshInterval, resetDelay, url, JSON.stringify(opts)]); // eslint-disable-line react-hooks/exhaustive-deps
-
-	useEffect(() => {
-		let timer;
-		if (timerSignal && resetDelay) {
-			timer = setTimeout(() => {
-				// reset the state
-				setData(null);
-				setIsFetching(false);
-				setIsFetched(false);
-				setError(null);
-			}, resetDelay);
-		}
-
-		return () => timer && clearTimeout(timer);
-	}, [resetDelay, timerSignal]);
-
-	useEffect(() => {
-		let timer;
-		if (timerSignal && refreshInterval) {
-			timer = setTimeout(fetchFn, refreshInterval);
-		}
-
-		return () => timer && clearTimeout(timer);
-	}, [fetchFn, refreshInterval, timerSignal]);
+	useRefreshInterval({ timerSignal, refreshInterval, fetchFn });
 
 	return {
 		isFetching,
